Move kite completion client to completion
