+++ spyder/plugins/completion/kite/decorators.py
