+++ spyder/plugins/completion/kite/plugin.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+# Copyright © Spyder Project Contributors
+# Licensed under the terms of the MIT License
+# (see spyder/__init__.py for details)
+
+"""Kite completions HTTP client."""
