+++ spyder/plugins/completion/kite/providers/document.py
